third_party/blink/web_tests/external/wpt/dom/nodes/MutationObserver-characterData-expected.txt

This is a testharness.js-based test.
[FAIL] characterData ProcessingInstruction: data mutations
  assert_equals: oldValue didn't match expected "?processing data?" but got "data"
Harness: the test ran to completion.
